# docs/api-specification.yaml
openapi: 3.1.0
info:
  title: AI Part Designer API
  description: |
    RESTful API for AI-powered 3D part design and generation platform.
    
    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the 
    `Authorization` header as `Bearer <token>`.
    
    ## Rate Limiting
    - Free tier: 30 requests/minute
    - Pro tier: 100 requests/minute
    - Enterprise tier: 500 requests/minute
    
    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in window
    - `X-RateLimit-Reset`: Unix timestamp when window resets
    
    ## Versioning
    API version is specified in the URL path: `/api/v1/`
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@aipartdesigner.com
  license:
    name: Proprietary
    url: https://aipartdesigner.com/terms
servers:
  - url: https://api.aipartdesigner.com/api/v1
    description: Production server
  - url: https://api.staging.aipartdesigner.com/api/v1
    description: Staging server
  - url: http://localhost:8000/api/v1
    description: Local development
...
paths:
...
  /templates:
    get:
      tags: [Templates]
      summary: List available templates
      operationId: listTemplates
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [mechanical, enclosures, connectors, hardware, organizational, decorative, custom]
        - name: tier
          in: query
          description: Filter by required tier (returns templates at or below this tier)
          schema:
            type: string
            enum: [free, pro, enterprise]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Templates list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Template'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /templates/{templateId}:
    get:
      tags: [Templates]
      summary: Get template details
      operationId: getTemplate
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Template details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /templates/{templateId}/preview:
    post:
      tags: [Templates]
      summary: Generate preview with custom parameters
      operationId: previewTemplate
      security:
        - bearerAuth: []
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                parameters:
                  type: object
                  additionalProperties: true
      responses:
        '200':
          description: Preview generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  previewUrl:
                    type: string
                    format: uri
                  validationErrors:
                    type: array
                    items:
                      type: object
                      properties:
                        parameter:
                          type: string
                        message:
                          type: string

  # ============================================================================
  # DESIGNS
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Template:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        category:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: string
          format: uri
        tierRequired:
          type: string
          enum: [free, pro, enterprise]
    TemplateDetail:
      allOf:
        - $ref: '#/components/schemas/Template'
        - type: object
          properties:
            parameters:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/ParameterDefinition'
            previewUrl:
              type: string
              format: uri
            usageCount:
              type: integer
    ParameterDefinition:
      type: object
      properties:
        type:
          type: string
          enum: [number, enum, boolean, string]
        label:
          type: string
        description:
          type: string
        default:
          oneOf:
            - type: number
            - type: string
            - type: boolean
        min:
          type: number
        max:
          type: number
        step:
          type: number
        unit:
          type: string
        options:
          type: array
          items:
            type: string
        required:
          type: boolean
          default: true

    # --------------------------------------------------------------------------
    # Designs
    # --------------------------------------------------------------------------
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            requestId:
              type: string
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
...
security:
  - bearerAuth: []